refactor(cli): move memory item CRUD under 'assistant memory items'
Move the item-level verbs out of the memory namespace root into a
dedicated 'items' subgroup (assistant memory items list|get|create|
update|delete), matching the namespace's existing subgroup layout
(v2, v3, retrospective, worker) and the oauth apps/providers
sibling-file pattern. Update help text, gateway risk-registry paths,
and tests accordingly.

Verified against a live daemon: create → update → get → delete under
the new path.
